fix(database): Ajusta tipos de colunas e adiciona role
- Altera o tipo das colunas de chave estrangeira para `unsignedBigInteger` nas migrations.
- Remove a migration `add_codigo_pessoa_to_users_table`.
- Adiciona a role `Operador` no `RoleSeeder`.

+++ database/migrations/2025_10_15_143818_create_grupos_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('grupos', function (Blueprint $table) {
             $table->id();
             $table->string('nome');
-            $table->integer('unidade_organizacional_id');
+            $table->unsignedBigInteger('unidade_organizacional_id');
             $table->foreign('unidade_organizacional_id')
                   ->references('id')
                   ->on('unidade_organizacionals')
